+++ changelog.md
@@ -12,6 +12,7 @@ Codion Change Log
 - State.Builder.group() now supports multiple groups.
 ### is.codion.common.rmi
 - AuxiliaryServerFactory.createServer() renamed create().
+- SerializationFilterFactory.createObjectInputFilter() renamed create().
 ### is.codion.swing.common.ui
 - MenuItemBuilder, unused type parameter removed.
 - Control.copy() return type fixed.